rozdzielone poszczególne klasy eksperymentów
