+++ PatternRecogniser/Models/Experiment.cs
@@ -12,12 +12,9 @@ namespace PatternRecogniser.Models
     {
         public ExtendedModel usedModel { get; set; }
 
-        public string GetResults()
-        {
-            return ""; // returns a string that follows json formatting
-        }
+        public abstract string GetResults();
 
-        public void SaveResult() { }
+        public abstract void SaveResult();
     }
 
     public class ExperimentList
@@ -26,24 +23,4 @@ namespace PatternRecogniser.Models
         public List<Experiment> experiments;
     }
 
-    public class ModelTrainingExperiment : Experiment
-    {
-        public List<(Bitmap, Pattern, Pattern)> validationSet;
-        public double accuracy { get; set; }
-        public double precision { get; set; }
-        public double recall { get; set; }
-        public double specificity { get; set; }
-        public double missRate { get; set; }
-        public int[,] confusionMatrix { get; set; }
-        private int TP;
-        private int TN;
-        private int FP;
-        private int FN;
-    }
-
-    public class PatternRecognitionExperiment : Experiment
-    {
-        public Bitmap testedPattern { get; set; }
-        public List<(Pattern, double)> recognisedPatterns;
-    }
 }
